Move FileMonitorSource instantiations out of monitor.cc
Bug: b/537428390
