+++ base/cvd/cuttlefish/host/commands/cvd/cli/commands/monitor/log_sources.h
@@ -1,0 +1,35 @@
+/*
+ * Copyright (C) 2026 The Android Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+#pragma once
+
+#include <memory>
+
+#include "cuttlefish/host/commands/cvd/cli/commands/monitor/monitor_source.h"
+#include "cuttlefish/host/commands/cvd/instances/local_instance.h"
+
+namespace cuttlefish {
+
+std::unique_ptr<MonitorSource> LauncherLogMonitorSource(
+    const LocalInstance& instance);
+
+std::unique_ptr<MonitorSource> KernelLogMonitorSource(
+    const LocalInstance& instance);
+
+std::unique_ptr<MonitorSource> LogcatMonitorSource(
+    const LocalInstance& instance);
+
+}  // namespace cuttlefish
