+++ base/cvd/cuttlefish/host/commands/cvd/cli/commands/monitor/monitor.cc
@@ -30,7 +30,6 @@
 #include <string>
 #include <string_view>
 #include <thread>
-#include <utility>
 #include <vector>
 
 #include "absl/cleanup/cleanup.h"
@@ -38,94 +37,15 @@
 
 #include "cuttlefish/ansi_codes/ansi_codes.h"
 #include "cuttlefish/common/libs/fs/shared_fd.h"
-#include "cuttlefish/files/file_exists.h"
 #include "cuttlefish/host/commands/cvd/cli/commands/monitor/display.h"
 #include "cuttlefish/host/commands/cvd/cli/commands/monitor/drain_inotify.h"
-#include "cuttlefish/host/commands/cvd/cli/commands/monitor/file_monitor_source.h"
-#include "cuttlefish/host/commands/cvd/cli/commands/monitor/kernel.h"
-#include "cuttlefish/host/commands/cvd/cli/commands/monitor/launcher.h"
-#include "cuttlefish/host/commands/cvd/cli/commands/monitor/log_tee.h"
-#include "cuttlefish/host/commands/cvd/cli/commands/monitor/logcat.h"
+#include "cuttlefish/host/commands/cvd/cli/commands/monitor/log_sources.h"
 #include "cuttlefish/host/commands/cvd/cli/commands/monitor/monitor_source.h"
 #include "cuttlefish/host/commands/cvd/cli/utils.h"
 #include "cuttlefish/host/commands/cvd/instances/local_instance.h"
-#include "cuttlefish/host/libs/log_names/log_names.h"
-#include "cuttlefish/io/io.h"
-#include "cuttlefish/io/shared_fd.h"
 #include "cuttlefish/result/result.h"
 
 namespace cuttlefish {
-namespace {
-
-Result<std::string> ColorLauncherOrLogTee(std::string_view line) {
-  if (line.find("log_tee(") != 0) {
-    return CF_EXPECT(ColorLauncherLine(line));
-  }
-  const size_t bracket = line.find(']');
-  if (bracket != std::string_view::npos) {
-    line.remove_prefix(bracket + 1);
-    const size_t first_non_space = line.find_first_not_of(" \t");
-    if (first_non_space != std::string_view::npos) {
-      line.remove_prefix(first_non_space);
-    } else {
-      line = "";
-    }
-  }
-  return CF_EXPECT(ColorLogTeeLine(line));
-}
-
-std::unique_ptr<MonitorSource> LauncherLogMonitorSource(
-    const LocalInstance& instance) {
-  const std::string launcher =
-      absl::StrCat(instance.InstanceDirectory(), "/logs/", kLogNameLauncher);
-  const std::string assemble =
-      absl::StrCat(instance.AssemblyDirectory(), "/", kLogNameAssembleCvd);
-  const std::string path = FileExists(launcher) ? launcher : assemble;
-  if (!FileExists(path)) {
-    return {};
-  }
-  SharedFD fd = SharedFD::Open(path, O_RDONLY);
-  if (!fd->IsOpen() || fd->LSeek(0, SEEK_END) <= 0) {
-    return {};
-  }
-  std::unique_ptr<ReaderSeeker> io = std::make_unique<SharedFdIo>(fd);
-  return std::make_unique<FileMonitorSource>(path, std::move(io),
-                                             ColorLauncherOrLogTee);
-}
-
-std::unique_ptr<MonitorSource> KernelLogMonitorSource(
-    const LocalInstance& instance) {
-  const std::string path =
-      absl::StrCat(instance.InstanceDirectory(), "/logs/", kLogNameKernel);
-  if (!FileExists(path)) {
-    return {};
-  }
-  SharedFD fd = SharedFD::Open(path, O_RDONLY);
-  if (!fd->IsOpen() || fd->LSeek(0, SEEK_END) <= 0) {
-    return {};
-  }
-  std::unique_ptr<ReaderSeeker> io = std::make_unique<SharedFdIo>(fd);
-  return std::make_unique<FileMonitorSource>(path, std::move(io),
-                                             ColorKernelLine);
-}
-
-std::unique_ptr<MonitorSource> LogcatMonitorSource(
-    const LocalInstance& instance) {
-  const std::string path =
-      absl::StrCat(instance.InstanceDirectory(), "/logs/", kLogNameLogcat);
-  if (!FileExists(path)) {
-    return {};
-  }
-  SharedFD fd = SharedFD::Open(path, O_RDONLY);
-  if (!fd->IsOpen() || fd->LSeek(0, SEEK_END) <= 0) {
-    return {};
-  }
-  std::unique_ptr<ReaderSeeker> io = std::make_unique<SharedFdIo>(fd);
-  return std::make_unique<FileMonitorSource>(path, std::move(io),
-                                             ColorLogcatLine);
-}
-
-}  // namespace
 
 Result<void> MonitorLogs(const LocalInstance& instance, SharedFD stop_eventfd) {
   CF_EXPECT(isatty(0), "The monitor command requires an interactive terminal.");
